feat(post): allow switch embed type of post

+++ src/components/Post/Form/PostEditForm.tsx
@@ -1,81 +1,71 @@
 import {FormikErrors, withFormik} from "formik"
 import PostForm, {PostFormValues} from "./PostForm"
-import {Embed, EmbedEdition, EmbedEnumType, EmbedGallery, EmbedMedia, EmbedPoll, EmbedPseudoGallery, Post} from "../../../data/post/types"
+import {Embed, EmbedCreation, EmbedEdition, EmbedEnumType, EmbedMedia, EmbedPoll, EmbedPseudoGallery, Post, PostUpdate} from "../../../data/post/types"
 import {mediaPath} from "../../../util"
 import {PostSizes} from "../../../constants/MediaSizes"
-import {updatePoll} from "../../../data/poll"
+import {createPoll, updatePoll} from "../../../data/poll"
 import {updatePost} from "../../../data/post"
 import {message} from "antd"
 import {createMedia} from "../../../data/media"
-import {addGalleryImages, deleteGalleryImages} from "../../../data/gallery"
+import {addGalleryImages, createGallery, deleteGalleryImages} from "../../../data/gallery"
+import {MediaUploadNSFW} from "../../../data/media/types"
 
 type PostEditFormProps = {
     post: Post,
     onEdit: (post: Post) => void
     onClose: () => void
 }
-const PostEditForm = withFormik<PostEditFormProps, PostFormValues<EmbedEdition>>({
+const PostEditForm = withFormik<PostEditFormProps, PostFormValues<EmbedEdition | EmbedCreation>>({
     mapPropsToValues: ({post}) => {
-        const embed: EmbedEdition = {
-            type: post.embed.embedType
-        } as EmbedEdition
+        let embed: EmbedEdition | undefined = undefined
+        if (post.embed) {
+            embed = {
+                type: post.embed.embedType
+            } as EmbedEdition
 
-        let postEmbed: Embed
-        switch (post.embed.embedType) {
-            case EmbedEnumType.GALLERY:
-                postEmbed = post.embed as EmbedGallery
-                if (postEmbed.pseudo) {
-                    embed.data = postEmbed.preview.map(img => ({
+            let postEmbed: Embed
+            switch (post.embed.embedType) {
+                case EmbedEnumType.IMAGE:
+                    postEmbed = post.embed as EmbedPseudoGallery
+                    embed.data = post.embed.images.map(img => ({
                         id: img.id,
                         file: mediaPath(img.name, PostSizes.PREVIEW) as string,
                         nsfw: img.nsfw
                     }))
-                } else {
-                    // Can't edit true gallery from post
-                    console.error("modification impossible")
-                }
-                break
-            case EmbedEnumType.IMAGE:
-                postEmbed = post.embed as EmbedPseudoGallery
-                embed.data = post.embed.images.map(img => ({
-                    id: img.id,
-                    file: mediaPath(img.name, PostSizes.PREVIEW) as string,
-                    nsfw: img.nsfw
-                }))
-                break
-            case EmbedEnumType.VIDEO:
-            case EmbedEnumType.DOCUMENT:
-                postEmbed = post.embed as EmbedMedia
-                embed.data = [{
-                    id: postEmbed.id,
-                    file: mediaPath(postEmbed.name, PostSizes.PREVIEW) as string,
-                    nsfw: postEmbed.nsfw
-                }]
-                break
-            case EmbedEnumType.POLL:
-                postEmbed = post.embed as EmbedPoll
-                embed.data = {
-                    ...postEmbed,
-                    choices: postEmbed.choices.map(c => ({
-                        id: c.id,
-                        content: c.content
-                    }))
-                }
-                break
+                    break
+                case EmbedEnumType.VIDEO:
+                case EmbedEnumType.DOCUMENT:
+                    postEmbed = post.embed as EmbedMedia
+                    embed.data = [{
+                        id: postEmbed.id,
+                        file: mediaPath(postEmbed.name, PostSizes.PREVIEW) as string,
+                        nsfw: postEmbed.nsfw
+                    }]
+                    break
+                case EmbedEnumType.POLL:
+                    postEmbed = post.embed as EmbedPoll
+                    embed.data = {
+                        ...postEmbed,
+                        choices: postEmbed.choices.map(c => ({
+                            id: c.id,
+                            content: c.content
+                        }))
+                    }
+                    break
+            }
         }
 
         return {
             id: post.id,
             description: post.description,
             publicationDate: post.publicationDate,
-            embed,
-            private: post.private,
-            draft: false
+            embed: embed,
+            private: post.private
         }
     },
 
     validate: (values) => {
-        const errors: FormikErrors<PostFormValues<EmbedEdition>> = {}
+        const errors: FormikErrors<PostFormValues<EmbedEdition | EmbedCreation>> = {}
         if (!values.description.length) {
             errors.description = "Required"
         }
@@ -83,45 +73,69 @@ const PostEditForm = withFormik<PostEditFormProps, PostFormValues<EmbedEdition>>
     },
 
     handleSubmit: async (values, {props}) => {
-        const {embed, ...post} = values
+        const {embed, ...postForm} = values
+        const post = postForm as PostUpdate
 
         if (embed) {
             switch (embed.type) {
                 case EmbedEnumType.IMAGE: {
-                    // Edit and create images
-                    const newImagesID = []
-                    for (const img of embed.data) {
-                        if ("id" in img) {
-                            console.log("oui")
-                        } else {
-                            const res = await createMedia(img)
-                            newImagesID.push(res.data.id)
+                    if (props.post.embed == null || embed.type != props.post.embed.embedType) {
+                        const ids = []
+                        for (const img of embed.data) {
+                            const res = await createMedia(img as MediaUploadNSFW)
+                            ids.push(res.data.id)
                         }
-                    }
-                    await addGalleryImages(props.post.embed.id, newImagesID)
 
-                    // Remove media that are no longer present in form
-                    const removeImagesID = []
-                    for (const img of (props.post.embed as EmbedPseudoGallery).images) {
-                        if (!embed.data.find(comparedImg => "id" in comparedImg && comparedImg.id == img.id)) {
-                            removeImagesID.push(img.id)
+                        const res = await createGallery({
+                            feed: 1, //TODO change this with real feed
+                            pseudo: true,
+                            images: ids
+                        })
+                        post.attachements = {[EmbedEnumType.IMAGE]: res.data.id}
+                    } else {
+                        // Edit and create images
+                        const newImagesID = []
+                        for (const img of embed.data) {
+                            if ("id" in img) {
+                                console.log("oui")
+                            } else {
+                                const res = await createMedia(img)
+                                newImagesID.push(res.data.id)
+                            }
                         }
-                    }
-                    await deleteGalleryImages(props.post.embed.id, removeImagesID)
+                        if (newImagesID.length > 0)
+                            await addGalleryImages(props.post.embed.id, newImagesID)
 
+                        // Remove media that are no longer present in form
+                        const removeImagesID = []
+                        for (const img of props.post.embed.images) {
+                            if (!embed.data.find(comparedImg => "id" in comparedImg && comparedImg.id == img.id)) {
+                                removeImagesID.push(img.id)
+                            }
+                        }
+                        if (removeImagesID.length > 0)
+                            await deleteGalleryImages(props.post.embed.id, removeImagesID)
+                    }
                     break
                 }
                 case EmbedEnumType.DOCUMENT:
                 case EmbedEnumType.VIDEO: {
-
+                    const media = embed.data[0]
+                    if (!("id" in media)) {
+                        const res = await createMedia(media)
+                        post.attachements = {[embed.type]: res.data.id}
+                    }
                     break
                 }
                 case EmbedEnumType.POLL: {
-                    await updatePoll(embed.data)
+                    if (props.post.embed == null || embed.type != props.post.embed.embedType) {
+                        const res = await createPoll(embed.data)
+                        post.attachements = {[EmbedEnumType.POLL]: res.data.id}
+                    } else {
+                        await updatePoll(embed.data)
+                    }
                     break
                 }
-                case EmbedEnumType.GALLERY:
-                default:
             }
         }
